+++ stacks/spica/packages/material/src/public_api.ts
@@ -3,3 +3,4 @@
  */
 export * from "./aware-dialog";
 export * from "./clipboard";
+export * from "./icons";
