+++ stacks/spica/src/passport/pages/strategies-add/strategies-add.component.ts
@@ -1,8 +1,9 @@
-import {Component, OnInit, ViewChild} from "@angular/core";
+import {Component, OnInit, ViewChild, OnDestroy} from "@angular/core";
+import {NgModel} from "@angular/forms";
 import {ActivatedRoute, Router} from "@angular/router";
+import {ICONS} from "@spica-client/material";
 import {Subject} from "rxjs";
 import {filter, switchMap, takeUntil} from "rxjs/operators";
-import {ICONS} from "src/bucket/icons";
 import {emptyStrategy, EMPTY_STRATEGY, Strategy} from "../../interfaces/strategy";
 import {StrategyService} from "../../strategy.service";
 
@@ -11,7 +12,7 @@ import {StrategyService} from "../../strategy.service";
   templateUrl: "./strategies-add.component.html",
   styleUrls: ["./strategies-add.component.scss"]
 })
-export class StrategiesAddComponent implements OnInit {
+export class StrategiesAddComponent implements OnInit, OnDestroy {
   @ViewChild("toolbar", {static: true}) toolbar;
 
   readonly icons: Array<string> = ICONS;
@@ -19,6 +20,7 @@ export class StrategiesAddComponent implements OnInit {
   public visibleIcons: Array<any> = this.icons.slice(0, this.iconPageSize);
 
   strategy: Strategy = emptyStrategy();
+  callbackUrl: string;
   private onDestroy: Subject<void> = new Subject<void>();
 
   constructor(
@@ -32,16 +34,26 @@ export class StrategiesAddComponent implements OnInit {
       .pipe(
         filter(params => params.id),
         takeUntil(this.onDestroy),
-        switchMap(params => this.strategyService.getStrategy(params.id))
+        switchMap(params => this.strategyService.getStrategy(params.id).toPromise())
       )
       .subscribe(strategyData => {
+        this.callbackUrl = strategyData.callbackUrl;
+        delete strategyData.callbackUrl;
         this.strategy = {...EMPTY_STRATEGY, ...strategyData};
       });
   }
 
-  submitForm() {
+  submitForm(certificate: NgModel) {
     this.strategyService
       .updateStrategy(this.strategy)
-      .then(() => this.router.navigate(["passport/strategies"]));
+      .toPromise()
+      .then(() => this.router.navigate(["passport/strategies"]))
+      .catch(err => {
+        certificate.control.setErrors({invalid: true});
+      });
+  }
+
+  ngOnDestroy() {
+    this.onDestroy.next();
   }
 }
